+++ utils/plasma-sb-utils/src/components/Select/index.ts
@@ -1,0 +1,1 @@
+export { getSelectStories } from './Select';
