+++ utils/plasma-sb-utils/src/components/index.ts
@@ -3,4 +3,6 @@ export { IconPlaceholder } from './IconPlaceholder';
 export { PaletteGrid } from './PaletteGrid';
 export { InSpacing as InSpacingDecorator } from './StoryDecorators';
 export { ThemeColors } from './ThemeColors';
+
 export * from './Button';
+export * from './Select';
